frontend: extract settings controllers
